frontend: common: refactored all imports to their direct location
