+++ frontend/src/components/App/Settings/DrawerModeSettings.tsx
@@ -21,7 +21,7 @@ import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { setDetailDrawerEnabled } from '../../../redux/drawerModeSlice';
 import { useTypedSelector } from '../../../redux/reducers/reducers';
-import { TooltipIcon } from '../../common';
+import { TooltipIcon } from '../../common/Tooltip';
 
 export default function DrawerModeSettings() {
   const dispatch = useDispatch();
