+++ frontend/src/components/DetailsViewSection/DetailsViewSection.stories.tsx
@@ -21,7 +21,7 @@ import { Provider } from 'react-redux';
 import { useDispatch } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import { makeMockKubeObject } from '../../test/mocker';
-import { SectionBox } from '../common';
+import { SectionBox } from '../common/SectionBox';
 import DetailsViewSection, { DetailsViewSectionProps } from './DetailsViewSection';
 import { setDetailsView } from './detailsViewSectionSlice';
