+++ frontend/src/components/authchooser/index.tsx
@@ -30,9 +30,10 @@ import { testAuth } from '../../lib/k8s/apiProxy';
 import { createRouteURL, getRoute, getRoutePath } from '../../lib/router';
 import { setConfig } from '../../redux/configSlice';
 import { ClusterDialog } from '../cluster/Chooser';
-import { Link, Loader } from '../common';
 import { DialogTitle } from '../common/Dialog';
 import Empty from '../common/EmptyContent';
+import Link from '../common/Link';
+import Loader from '../common/Loader';
 import OauthPopup from '../oidcauth/OauthPopup';
 
 const ColorButton = styled(Button)(({ theme }) => ({
