+++ frontend/src/components/cluster/Overview.tsx
@@ -27,7 +27,10 @@ import Pod from '../../lib/k8s/pod';
 import { useFilterFunc } from '../../lib/util';
 import { OverviewChart } from '../../redux/overviewChartsSlice';
 import { useTypedSelector } from '../../redux/reducers/reducers';
-import { DateLabel, Link, PageGrid, StatusLabel } from '../common';
+import { DateLabel } from '../common/Label';
+import { StatusLabel } from '../common/Label';
+import Link from '../common/Link';
+import { PageGrid } from '../common/Resource';
 import ResourceListView from '../common/Resource/ResourceListView';
 import { SectionBox } from '../common/SectionBox';
 import ShowHideLabel from '../common/ShowHideLabel';
