+++ frontend/src/components/common/ObjectEventList.tsx
@@ -20,7 +20,9 @@ import Event, { KubeEvent } from '../../lib/k8s/event';
 import { KubeObject } from '../../lib/k8s/KubeObject';
 import { localeDate, timeAgo } from '../../lib/util';
 import { HeadlampEventType, useEventCallback } from '../../redux/headlampEventSlice';
-import { HoverInfoLabel, SectionBox, SimpleTable } from '../common';
+import { HoverInfoLabel } from '../common/Label';
+import SectionBox from '../common/SectionBox';
+import SimpleTable from '../common/SimpleTable';
 import ShowHideLabel from './ShowHideLabel';
 
 export interface ObjectEventListProps {
