+++ frontend/src/components/crd/Details.tsx
@@ -18,8 +18,9 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import CRD from '../../lib/k8s/crd';
-import { Link, ObjectEventList } from '../common';
+import Link from '../common/Link';
 import Loader from '../common/Loader';
+import ObjectEventList from '../common/ObjectEventList';
 import { ConditionsTable, MainInfoSection, PageGrid } from '../common/Resource';
 import { SectionBox } from '../common/SectionBox';
 import SimpleTable from '../common/SimpleTable';
