+++ frontend/src/components/crd/List.tsx
@@ -18,7 +18,8 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import CRD from '../../lib/k8s/crd';
 import { useNamespaces } from '../../redux/filterSlice';
-import { Link, useThrottle } from '../common';
+import Link from '../common/Link';
+import { useThrottle } from '../common/Resource';
 import ResourceListView from '../common/Resource/ResourceListView';
 
 export default function CustomResourceDefinitionList() {
