+++ frontend/src/components/limitRange/List.tsx
@@ -18,8 +18,8 @@ import { useTranslation } from 'react-i18next';
 import { ApiError } from '../../lib/k8s/apiProxy';
 import { LimitRange } from '../../lib/k8s/limitRange';
 import { useNamespaces } from '../../redux/filterSlice';
-import { SimpleTableProps } from '../common';
 import ResourceListView from '../common/Resource/ResourceListView';
+import { SimpleTableProps } from '../common/SimpleTable';
 
 export interface LimitRangeProps {
   limitRanges: LimitRange[] | null;
