+++ frontend/src/components/namespace/CreateNamespaceButton.tsx
@@ -29,7 +29,8 @@ import Namespace from '../../lib/k8s/namespace';
 import { clusterAction } from '../../redux/clusterActionSlice';
 import { EventStatus, HeadlampEventType, useEventCallback } from '../../redux/headlampEventSlice';
 import { AppDispatch } from '../../redux/stores/store';
-import { ActionButton, AuthVisible } from '../common';
+import ActionButton from '../common/ActionButton';
+import { AuthVisible } from '../common/Resource';
 
 export default function CreateNamespaceButton() {
   const { t } = useTranslation(['glossary', 'translation']);
