+++ frontend/src/components/node/Details.tsx
@@ -32,7 +32,9 @@ import { DefaultHeaderAction } from '../../redux/actionButtonsSlice';
 import { clusterAction } from '../../redux/clusterActionSlice';
 import { AppDispatch } from '../../redux/stores/store';
 import { CpuCircularChart, MemoryCircularChart } from '../cluster/Charts';
-import { ActionButton, ConfirmDialog, StatusLabelProps } from '../common';
+import ActionButton from '../common/ActionButton';
+import ConfirmDialog from '../common/ConfirmDialog';
+import { StatusLabelProps } from '../common/Label';
 import { HeaderLabel, StatusLabel, ValueLabel } from '../common/Label';
 import { ConditionsSection, DetailsGrid, OwnedPodsSection } from '../common/Resource';
 import AuthVisible from '../common/Resource/AuthVisible';
