+++ frontend/src/components/resourceQuota/Details.tsx
@@ -18,7 +18,8 @@ import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import ResourceQuota from '../../lib/k8s/resourceQuota';
 import { compareUnits, normalizeUnit } from '../../lib/util';
-import { DetailsGrid, SimpleTable } from '../common';
+import { DetailsGrid } from '../common/Resource';
+import SimpleTable from '../common/SimpleTable';
 
 export default function ResourceQuotaDetails(props: {
   name?: string;
