+++ frontend/src/components/service/Details.tsx
@@ -22,9 +22,9 @@ import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import Endpoint from '../../lib/k8s/endpoints';
 import Service from '../../lib/k8s/service';
-import { Link } from '../common';
 import Empty from '../common/EmptyContent';
 import { ValueLabel } from '../common/Label';
+import Link from '../common/Link';
 import { DetailsGrid, MetadataDictGrid } from '../common/Resource';
 import PortForward from '../common/Resource/PortForward';
 import { SectionBox } from '../common/SectionBox';
