+++ frontend/src/components/serviceaccount/Details.tsx
@@ -18,7 +18,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import ServiceAccount from '../../lib/k8s/serviceAccount';
-import { Link } from '../common';
+import Link from '../common/Link';
 import { DetailsGrid } from '../common/Resource';
 
 export default function ServiceAccountDetails(props: {
